toolpool/Materials/Basic_Functionality/requirement.txt
fastmindapi
pickle
typing
pymatgen.core
pymatgen.symmetry.analyzer
pymatgen.analysis.structure_matcher
pymatgen.io.vasp.sets